Rename classes and copies for top and hot photos scrteens
