+++ app/src/main/java/com/jp/imggram/ui/dashboard/HotPhotosViewModel.kt
@@ -1,13 +1,13 @@
-package com.jp.imggram.ui.notifications
+package com.jp.imggram.ui.dashboard
 
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 
-class NotificationsViewModel : ViewModel() {
+class HotPhotosViewModel : ViewModel() {
 
     private val _text = MutableLiveData<String>().apply {
-        value = "This is notifications Fragment"
+        value = "This is hot photos Fragment"
     }
     val text: LiveData<String> = _text
 }
