+++ app/src/main/java/com/jp/imggram/ui/home/TopPhotosViewModel.kt
@@ -1,13 +1,13 @@
-package com.jp.imggram.ui.dashboard
+package com.jp.imggram.ui.home
 
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 
-class DashboardViewModel : ViewModel() {
+class TopPhotosViewModel : ViewModel() {
 
     private val _text = MutableLiveData<String>().apply {
-        value = "This is dashboard Fragment"
+        value = "This is home Fragment"
     }
     val text: LiveData<String> = _text
 }
