+++ app/src/main/res/layout/fragment_hot_photos.xml
@@ -4,10 +4,10 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ui.notifications.NotificationsFragment">
+    tools:context=".ui.dashboard.HotPhotosFragment">
 
     <TextView
-        android:id="@+id/text_notifications"
+        android:id="@+id/text_photos"
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
         android:layout_marginStart="8dp"
